QcomModulePkg: Update the type of GetTimerCountms
Use UINT64 instead of UINT32 to fix the potential data loss.

+++ QcomModulePkg/Include/Library/LinuxLoaderLib.h
@@ -153,7 +153,7 @@ GetBlkIOHandles (
                  );
 
 VOID ToLower(CHAR8 *Str);
-UINT32 GetTimerCountms (VOID);
+UINT64 GetTimerCountms (VOID);
 EFI_STATUS WriteToPartition (EFI_GUID *Ptype, VOID *Msg, UINT32 MsgSize);
 BOOLEAN IsSecureBootEnabled(VOID);
 EFI_STATUS ResetDeviceState(VOID);
